extract time entry modals into common base class
