+++ frontend/src/app/modules/time_entries/shared/modal/base.modal.html
@@ -22,20 +22,32 @@
 
     <div class="op-modal--modal-footer">
       <button class="button -highlight"
-              (click)="createEntry()"
-              [attr.title]="text.create">
+              *ngIf="saveAllowed"
+              (click)="saveEntry()"
+              [attr.title]="saveText">
         <op-icon icon-classes="button--icon icon-checkmark"></op-icon>
         <span class="button--text"
-              [textContent]="text.create"
+              [textContent]="saveText"
               aria-hidden="true"></span>
       </button>
 
       <button class="button"
+              *ngIf="saveAllowed"
               (click)="closeMe($event)"
               [attr.title]="text.cancel">
         <op-icon icon-classes="button--icon icon-close"></op-icon>
         <span class="button--text"
               [textContent]="text.cancel"
+              aria-hidden="true"></span>
+      </button>
+
+      <button class="button"
+              *ngIf="deleteAllowed"
+              (click)="destroy()"
+              [attr.title]="text.delete">
+        <op-icon icon-classes="button--icon icon-delete"></op-icon>
+        <span class="button--text"
+              [textContent]="text.delete"
               aria-hidden="true"></span>
       </button>
     </div>
