+++ frontend/src/app/modules/time_entries/shared/modal/base.modal.sass
